fundExpert/pyproject.toml
[project]
name = "fundexpert"
version = "0.1.0"
description = "基于 CrewAI 的全天候策略基金分析服务"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.1.0",
    "crewai-tools>=0.1.0",
    "akshare>=1.12.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "api", "tools"]
